Fix google map and Sign out

+++ app/src/main/java/com/example/ken/updish/BackgroundWorker/PostDetailsBackgroundWorker.java
@@ -175,8 +175,8 @@ public class PostDetailsBackgroundWorker extends AsyncTask<String, Void, String>
             loc.setAddress(obj.getString("address"));
 
             // Ken
-            loc.setLatitude(obj.getLong("latitude"));
-            loc.setLongtitude(obj.getLong("longtitude"));
+            loc.setLatitude(obj.getDouble("latitude"));
+            loc.setLongtitude(obj.getDouble("longtitude"));
 
             tempPost.setLocation(loc);
